made keep-alive script exitable
The startup script that kept the client application running in an
infinite loop is now terminable under certain conditions. It monitors
the exit code of the skolar-display client application. If the exit code
has a certain value (42), it will not restart the application again.

This is mostly a convenience function for testing and debugging
purposes.